requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.47.0
django-countries==7.2.1
django-environ==0.8.1
django-iban==0.3.1
django-localflavor==3.1
executing==0.8.2
idna==3.3
ipython==8.0.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
oauthlib==3.1.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python-decouple==3.5
python-stdnum==1.17
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==1.2.3
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
